requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
mypy>=1.5.0
isort>=5.12.0
pylint>=2.17.0
bandit[toml]>=1.7.5

# Type stubs
boto3-stubs[essential]>=1.28.0
types-requests>=2.31.0
types-python-dateutil>=2.8.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0

# Development tools
pre-commit>=3.3.0
ipython>=8.14.0
ipdb>=0.13.13

# Performance profiling
memory-profiler>=0.61.0
line-profiler>=4.1.0